+++ backend/app/observability/metrics.py
@@ -18,17 +18,29 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
+import logging
 
+import httpx
+import yaml
 from opentelemetry import metrics
+from opentelemetry.metrics import Observation
 from opentelemetry.sdk.metrics import (
-    MeterProvider, Counter, UpDownCounter, Histogram
+    MeterProvider,
+    Counter,
+    UpDownCounter,
+    Histogram,
 )
 from opentelemetry.sdk.metrics.export import (
-    PeriodicExportingMetricReader, AggregationTemporality
+    PeriodicExportingMetricReader,
+    AggregationTemporality,
 )
 from opentelemetry.exporter.otlp.proto.grpc.metric_exporter import (
     OTLPMetricExporter,
 )
+from pydantic import BaseModel, ConfigDict
+from pydantic.alias_generators import to_camel
+
+from app.settings import settings
 
 metric_exporter = OTLPMetricExporter(
     preferred_temporality={
@@ -68,4 +80,148 @@ session_duration = meter.create_histogram(
     "sessions.duration",
     unit="seconds",
     description="Session duration in seconds.",
+)
+
+
+class ClusterQuota(BaseModel):
+    names: str
+    limit: int
+
+
+class GpuQuota(BaseModel):
+    name: str
+    clusters: list[ClusterQuota] | None = None
+
+
+class GpuQuotaResponse(BaseModel):
+    gpus: list[GpuQuota]
+
+
+def get_total_gpus(options):
+    if not settings.ngc_endpoint or not settings.ngc_org:
+        return
+
+    with httpx.Client() as client:
+        try:
+            response = client.get(
+                f"{settings.ngc_endpoint}/v2/orgs/{settings.ngc_org}/ngc/nvcf/gpu/quota/rules",
+                headers={
+                    "Authorization": f"Bearer {settings.nvcf_api_key}",
+                },
+                timeout=15,
+            )
+            if response.status_code != 200:
+                logging.error(
+                    f"Failed to get GPU information: HTTP{response.status_code} -- "
+                    f"{response.text}"
+                )
+                return
+        except TimeoutError:
+            logging.debug(f"Failed to get GPU information: timeout.")
+            return
+
+        body = yaml.safe_load(response.text)
+        quota = GpuQuotaResponse.model_validate(body)
+
+        for gpu in quota.gpus:
+            if gpu.clusters is not None:
+                for cluster in gpu.clusters:
+                    yield Observation(
+                        cluster.limit,
+                        {"cluster": cluster.names, "gpu": gpu.name},
+                    )
+
+
+gpu_total = meter.create_observable_gauge(
+    "gpu.total.count",
+    unit="gpus",
+    description="Total number of GPUs available for the NGC organization.",
+    callbacks=[get_total_gpus],
+)
+
+
+class GpuClusterUsage(BaseModel):
+    min_instances: int
+    max_instances: int
+    active_instances: int
+    available_instances: int
+    active_gpus: int
+    available_gpus: int
+
+    model_config = ConfigDict(populate_by_name=True, alias_generator=to_camel)
+
+
+class GpuCluster(BaseModel):
+    cluster_id: str
+    cluster_name: str
+    usage: GpuClusterUsage
+
+    model_config = ConfigDict(populate_by_name=True, alias_generator=to_camel)
+
+
+class GpuInstanceRegion(BaseModel):
+    region_name: str
+    clusters: list[GpuCluster]
+
+    model_config = ConfigDict(populate_by_name=True, alias_generator=to_camel)
+
+
+class GpuInstance(BaseModel):
+    instance_name: str
+    value: str
+    regions: list[GpuInstanceRegion]
+
+    model_config = ConfigDict(populate_by_name=True, alias_generator=to_camel)
+
+
+class InstanceTypeResponse(BaseModel):
+    gpus: dict[str, list[GpuInstance]]
+
+
+def get_active_gpus(options):
+    if not settings.ngc_endpoint or not settings.ngc_org:
+        return
+
+    with httpx.Client() as client:
+        try:
+            response = client.get(
+                f"{settings.ngc_endpoint}/v3/orgs/{settings.ngc_org}/ngc/nvcf/deployments/instanceTypes",
+                headers={
+                    "Authorization": f"Bearer {settings.nvcf_api_key}",
+                },
+                timeout=15,
+            )
+            if response.status_code != 200:
+                logging.error(
+                    f"Failed to get active gpus: HTTP{response.status_code} -- "
+                    f"{response.text}"
+                )
+                return
+        except TimeoutError:
+            logging.debug(f"Failed to get active gpus: timeout.")
+            return
+
+        body = response.json()
+        instances = InstanceTypeResponse.model_validate({"gpus": body})
+
+        for gpu, instances in instances.gpus.items():
+            for instance in instances:
+                for region in instance.regions:
+                    for cluster in region.clusters:
+                        yield Observation(
+                            cluster.usage.active_gpus,
+                            {
+                                "gpu": gpu,
+                                "instance": instance.instance_name,
+                                "region": region.region_name,
+                                "cluster": cluster.cluster_name,
+                            },
+                        )
+
+
+gpu_active = meter.create_observable_gauge(
+    "gpu.active.count",
+    unit="gpus",
+    description="GPUs currently assigned to active NVCF functions.",
+    callbacks=[get_active_gpus],
 )